Shape/Scalar Correlation

+++ Libs/Analyze/Analyze.h
@@ -57,7 +57,7 @@ class Analyze {
 
   ShapeHandle create_shape_from_points(Particles points);
 
-  Eigen::VectorXf get_subject_features(int subject, std::string feature_name);
+  Eigen::VectorXd get_subject_features(int subject, std::string feature_name);
 
   void set_group_selection(std::string group_name, std::string group1, std::string group2);
