+++ Studio/Job/ParticleAreaJob.h
@@ -17,14 +17,14 @@ class ParticleAreaJob : public Job {
   void run() override;
   QString name() override { return "Particle Normal Evaluation"; }
 
-  const Eigen::VectorXf& get_mean_areas() const { return mean_areas_; }
-  const Eigen::VectorXf& get_stddev_areas() const { return stddev_areas_; }
+  const Eigen::VectorXd& get_mean_areas() const { return mean_areas_; }
+  const Eigen::VectorXd& get_stddev_areas() const { return stddev_areas_; }
 
  Q_SIGNALS:
 
  private:
   QSharedPointer<Session> session_;
-  Eigen::VectorXf mean_areas_;
-  Eigen::VectorXf stddev_areas_;
+  Eigen::VectorXd mean_areas_;
+  Eigen::VectorXd stddev_areas_;
 };
 }  // namespace shapeworks
